cf_verification/requirements.txt
functions-framework==3.*
google-cloud-logging
google-cloud-firestore
google-cloud-tasks==2.7.1
google-python-cloud-debugger
requests
pydantic
cachetools
aiohttp[speedups]
aiocache
python-dotenv
